// src/interfaces/DataConfigProfileInterface.ts
export interface DataConfigProfileInterface {
  nombre?: string
  apellido?: string
  cliente?: string
  proveed?: string
  rifci?: string
  rif?: string
  cedula?: string
  email?: string
  emailc?: string
  telefono?: string
  aniversario?: string
  nacimi?: string
  dire11?: string
  direc1?: string
  us_nombre?: string
  telefon2?: string
  contacto?: string
}